refactor: simplify hwygen transformer and IR analysis
Reduce code duplication and fix bugs across the hwygen code generator:

transformer.go (-492 lines net):
- Extract unsafeSlicePointer helper, replacing 12+ duplicate AST constructions
- Extract getHwygenTypeSuffix helper, deduplicating type-suffix logic
- Extract classifyTypeParams helper from 3 near-identical call sites
- Merge duplicate SlideUpLanes/SlideDownLanes and AllTrue/AllFalse switch cases
- Replace isLikelyConstant heuristic with deterministic PackageConsts lookup
- Fix collectDeclaredVars to use ast.Inspect instead of manual walk

c_ast_translator.go (bug fix):
- Fix int64/uint64 C type mappings: use "long long"/"unsigned long long"
  instead of "long"/"unsigned long" to match 64-bit width on all platforms

ir/analysis.go (-159 lines net):
- Remove dead defInfo.uses field and unused walkUses traversal
- Delete dead exported functions ComputeDepthFirst, ComputeTopologicalOrder
- Fix IdentifyCrossLoopAllocations to find nested loops via WalkNodes
- Merge double child-scan into single pass in IdentifyCrossLoopAllocations
- Add early exit to canEliminateAlloc

ir/types.go, ir/fusion.go, ir/emitter.go:
- Add WalkNodes helper, replacing 7 duplicate recursive closures
- Add pattern constants (PatternElemElem, etc.), replacing 15 string literals

+++ cmd/hwygen/c_math_helpers.go
@@ -891,7 +891,7 @@ var neonU64MaskQueryHelpers = []string{
 }`,
 	`static inline uint64x2_t hwy_first_n_u64(long n) {
     uint64x2_t iota = {0, 1};
-    return vcltq_u64(iota, vdupq_n_u64((unsigned long)n));
+    return vcltq_u64(iota, vdupq_n_u64((unsigned long long)n));
 }`,
 }
 
@@ -1003,14 +1003,14 @@ var neonU64MaxMinHelpers = []string{
     uint64x2_t mask = vcltq_u64(a, b);
     return vbslq_u64(mask, a, b);
 }`,
-	`static inline unsigned long hwy_reducemax_u64(uint64x2_t v) {
-    unsigned long a = vgetq_lane_u64(v, 0);
-    unsigned long b = vgetq_lane_u64(v, 1);
+	`static inline unsigned long long hwy_reducemax_u64(uint64x2_t v) {
+    unsigned long long a = vgetq_lane_u64(v, 0);
+    unsigned long long b = vgetq_lane_u64(v, 1);
     return a > b ? a : b;
 }`,
-	`static inline unsigned long hwy_reducemin_u64(uint64x2_t v) {
-    unsigned long a = vgetq_lane_u64(v, 0);
-    unsigned long b = vgetq_lane_u64(v, 1);
+	`static inline unsigned long long hwy_reducemin_u64(uint64x2_t v) {
+    unsigned long long a = vgetq_lane_u64(v, 0);
+    unsigned long long b = vgetq_lane_u64(v, 1);
     return a < b ? a : b;
 }`,
 }
@@ -1024,7 +1024,7 @@ var neonS64MaskHelpers = slices.Concat(neonU64MaskQueryHelpers, []string{
     int64x2_t v = {0, 1};
     return v;
 }`,
-	`static inline long hwy_compress_store_s64(int64x2_t v, uint64x2_t mask, long *dst) {
+	`static inline long hwy_compress_store_s64(int64x2_t v, uint64x2_t mask, long long *dst) {
     long count = 0;
     if (vgetq_lane_u64(mask, 0)) { dst[count++] = vgetq_lane_s64(v, 0); }
     if (vgetq_lane_u64(mask, 1)) { dst[count++] = vgetq_lane_s64(v, 1); }
@@ -1041,7 +1041,7 @@ var neonU64MaskHelpers = slices.Concat(neonU64MaskQueryHelpers, []string{
     uint64x2_t v = {0, 1};
     return v;
 }`,
-	`static inline long hwy_compress_store_u64(uint64x2_t v, uint64x2_t mask, unsigned long *dst) {
+	`static inline long hwy_compress_store_u64(uint64x2_t v, uint64x2_t mask, unsigned long long *dst) {
     long count = 0;
     if (vgetq_lane_u64(mask, 0)) { dst[count++] = vgetq_lane_u64(v, 0); }
     if (vgetq_lane_u64(mask, 1)) { dst[count++] = vgetq_lane_u64(v, 1); }
